Move create obs and create with no evidence into Observation model
